ENH: Restore original register contents.
The plugin may modify the register contents, e.g. change "foo, " to ", foo". This implementation detail should not be visible to the user, so that he is able to insert the original register contents (e.g. with i_CTRL-R) after a smart put.
This patch appends a <sid>restore mapping that restores the saved original register contents after the modified <sid>put.